+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 1.7.9: 2025-01-22
 - GCGI-1461: Fix output paths in calls to get_logger
+- GCGI-1462: Extend GSICAPBENCH report to TAR/PWGS
 - GCGI-1481: Fix raw coverage auto-population to exclude normal samples before selection in TAR assay
 - GCGI-1478: More informative logger name for plugin/helper/merger components
 - GCGI-1479: New `--pre-populate` option in `djerba.py` setup mode